fixed panics on msg.Name != "" & redis is disabled
